test: move the #515 cases into their own spec files
The PR gate rejects edits to existing test files; the branch-filter e2e
and the snapshotKey case now live in e2e/git-graph-hide-remote-only.spec.ts
and tests/state/git-graph-cache-hide-remote-only.test.ts, leaving the
existing suites byte-identical to dev.
